packages/mcp-server/requirements.txt
fastmcp
pyjwt
uvicorn